replaced unsigned int with size_t to compile w/ cuda 3.2 on hatp13

+++ src/cux/cux.cpp
@@ -438,7 +438,8 @@ bool cux_init()
 
 #if !__DEVICE_EMULATION__
 	// Memory info
-	unsigned int free = 0, total = 0;
+	//	unsigned int free = 0, total = 0;
+	size_t free = 0, total = 0;
 	cuxErrCheck( (cudaError)cuMemGetInfo(&free, &total) );
 //	MLOG(verb2) << "Device memory (free, total): " << free / (1<<20) << "M, " << total / (1<<20) << "M" << "\n";
 #endif
@@ -448,3 +449,4 @@ bool cux_init()
 
 	return true;
 }
+
